# public/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AppDirect AI API",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Authorization",
        "description": "Enter your bearer token in the format **Bearer {token}**"
      }
    },
    "schemas": {
...
      "ChatSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the chat session."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the chat session was created."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the chat session was last updated."
          },
          "name": {
            "type": "string",
            "description": "Name of the chat session."
          },
          "userId": {
            "type": "string",
            "description": "Identifier of the user associated with the chat session."
          },
          "pinPosition": {
            "type": "integer",
            "format": "int32",
            "description": "The position of the chat in a pinned list or similar.",
            "nullable": true
          },
          "ai": {
            "$ref": "#/components/schemas/ChatAI"
          }
        }
      },
      "ChatDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChatSummary"
          },
          {
            "type": "object",
            "properties": {
              "messages": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ChatMessage"
                }
              }
            }
          }
        ]
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the chat message."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the chat message was created."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the chat message was last updated."
          },
          "content": {
            "type": "string",
            "description": "The content of the message"
          },
          "role": {
            "type": "string",
            "description": "The role of the message sender, either user or system"
          }
        }
      },
      "ChatAI": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the AI."
          },
          "name": {
            "type": "string",
            "description": "Name of the AI."
          },
          "src": {
            "type": "string",
            "description": "URL for the AI's image."
          },
          "description": {
            "type": "string",
            "description": "Description of the AI."
          },
          "userId": {
            "type": "string",
            "description": "Identifier of the AI's owner."
          },
          "userName": {
            "type": "string",
            "description": "The user name of the AI's owner."
          }
        }
      },
...
      "StreamingTextResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The streaming text response from the AI."
          }
        }
      },
...
      "GroupAvailability": {
        "type": "string",
        "enum": ["EVERYONE", "RESTRICTED"],
        "description": "Availability status of the group."
      },
      "GroupSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "orgId": {
            "type": "string"
          },
          "ownerUserId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "availability": {
            "$ref": "#/components/schemas/GroupAvailability"
          }
        }
      },
      "GroupDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/GroupSummary"
          },
          {
            "type": "object",
            "properties": {
              "users": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/GroupUser"
                }
              }
            }
          }
        ]
      },
      "GroupUser": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/v1/chats/{chatId}": {
      "get": {
        "summary": "Get a chat session",
        "description": "Retrieves the chat session with the specified ID.",
        "operationId": "getChat",
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the chat session to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat session successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden, the user is not authorized to perform this action."
          },
          "404": {
            "description": "Not found, when the specified chat ID does not exist."
          },
          "500": {
            "description": "Internal Server Error, any internal error."
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "summary": "Start a chat session",
        "description": "Initiates a chat with the AI using the provided prompt on the specified date.",
        "operationId": "postChatSession",
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the chat session.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "The date of the chat."
                  },
                  "prompt": {
                    "type": "string",
                    "description": "The prompt to use for chatting with an AI."
                  }
                },
                "required": ["date", "prompt"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a streaming text response from the AI.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamingTextResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, when the request body does not contain the required fields or contains invalid data."
          },
          "403": {
            "description": "Forbidden, the user is not authorized to perform this action."
          },
          "404": {
            "description": "Not found, when the specified chat ID does not exist."
          },
          "500": {
            "description": "Internal Server Error, any internal error."
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a chat session",
        "description": "Deletes the chat session with the specified ID.",
        "operationId": "deleteChat",
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the chat session to delete.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Chat session successfully deleted, no content to return."
          },
          "403": {
            "description": "Forbidden, the user is not authorized to perform this action."
          },
          "404": {
            "description": "Not Found, the specified chat ID does not exist."
          },
          "500": {
            "description": "Internal Server Error, any internal error."
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/groups/{groupId}": {
      "put": {
        "summary": "Update a group",
        "description": "Updates the details of a specific group and manages its member list.",
        "operationId": "updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the group to update.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the group."
                  },
                  "availability": {
                    "$ref": "#/components/schemas/GroupAvailability"
                  },
                  "memberEmailsToAdd": {
                    "type": "string",
                    "description": "A comma-separated list of emails of users to be added to the group."
                  },
                  "memberEmailsToRemove": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "A comma-separated list of emails of users to be removed from the group."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group details updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, when the input data is invalid."
          },
          "403": {
            "description": "Forbidden, the user does not have permission to update the group."
          },
          "404": {
            "description": "Not found, when the specified group ID does not exist."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "summary": "Delete a group",
        "description": "Deletes the group with the specified identifier.",
        "operationId": "deleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the group to delete.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Group successfully deleted, no content to return."
          },
          "400": {
            "description": "Bad request, when the request is malformed."
          },
          "403": {
            "description": "Forbidden, the user does not have permission to delete the group."
          },
          "404": {
            "description": "Not found, when the specified group ID does not exist."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  },
...
}